login navigation complete
