+++ package.json
@@ -10,11 +10,18 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "@react-native-community/masked-view": "^0.1.10",
     "@react-native-firebase/app": "^7.2.0",
     "@react-native-firebase/auth": "^8.0.4",
     "@react-native-firebase/firestore": "^7.1.7",
+    "@react-navigation/native": "^5.5.1",
+    "@react-navigation/stack": "^5.5.1",
     "react": "16.11.0",
     "react-native": "0.62.2",
+    "react-native-gesture-handler": "^1.6.1",
+    "react-native-reanimated": "^1.9.0",
+    "react-native-safe-area-context": "^3.0.5",
+    "react-native-screens": "^2.8.0",
     "watchman": "^1.0.0"
   },
   "devDependencies": {
